chore: restructure as 'src layout' Python package
