+++ src/fmask/bitlib.py
