+++ src/fmask/lightgbmlib.py
@@ -7,9 +7,9 @@ import pickle
 from pathlib import Path
 import numpy as np
 import pandas as pd
-import constant as C
+from . import constant as C
 import lightgbm as lgb
-from utils import collect_sample_pixel
+from .utils import collect_sample_pixel
 np.seterr(invalid='ignore') # ignore the invalid errors
 
 class Dataset:
